Wip on adding datasourceV2 compatibility
