+++ src/main/scala/io/qbeast/spark/internal/sources/v2/QbeastWriteBuilder.scala
@@ -1,0 +1,5 @@
+package io.qbeast.spark.internal.sources.v2
+
+import org.apache.spark.sql.connector.write.WriteBuilder
+
+class QbeastWriteBuilder extends WriteBuilder {}
