+++ src/test/scala/io/qbeast/spark/utils/QbeastSQLIntegrationTest.scala
@@ -1,0 +1,5 @@
+package io.qbeast.spark.utils
+
+import io.qbeast.spark.QbeastIntegrationTestSpec
+
+class QbeastSQLIntegrationTest extends QbeastIntegrationTestSpec {}
